+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - season - minus operator
 - update season (adding to season_from, _to)
 - spec rake: update:grandprix
+- grandprix: dont run if FINAL RESULTS ready
 
 ### new feature